+++ components/Drawer/types.ts
@@ -1,5 +1,6 @@
 export interface IDrawerPropTypes extends React.PropsWithChildren {
   className?: string;
+  keyboard?: boolean;
   maskClosable?: boolean;
   onClose?: () => void;
   open?: boolean;