+++ components/Modal/modal.tsx
@@ -6,6 +6,7 @@ import type { IModalPropTypes } from "./types";
 const Modal: React.FC<IModalPropTypes> = (props) => {
   const {
     children,
+    keyboard = true,
     maskClosable = true,
     onClose,
     open,
@@ -38,11 +39,20 @@ const Modal: React.FC<IModalPropTypes> = (props) => {
       }, 200);
     }
 
+    const handleEscKey = (event: any) => {
+      if (event.keyCode === 27) {
+        onClose?.();
+      }
+    };
+
+    if (keyboard) document.addEventListener("keydown", handleEscKey);
+
     return () => {
       if (!open) {
         openTimer && clearTimeout(openTimer);
         leaveTimer && clearTimeout(leaveTimer);
       }
+      if (keyboard) document.removeEventListener("keydown", handleEscKey);
     };
   }, [open]);